Avanços no movimento do veículo O movimento do tanque está exatamente igual ao movimento antigo em C (só que em inglês). Espero deixar o movimento mais realista, talvez separar entre movimento de tanque e carro (a diferença está no eixo de rotação). Também criei um arquivo para as funções relacionadas a física.

+++ src/FrameManager.hpp
@@ -14,7 +14,7 @@ class FrameManager {
 		static void init(int targetFPS);		
 		static bool frameDue();
 		
-		static int getDeltaTime(); 
+		static float getDeltaTime(); 
 };
 
 #endif
